refactor: rename potential interop
