+++ src/galax/_interop/matplotlib/__init__.py
@@ -1,5 +1,9 @@
 """Matplotlib extension for :mod:`galax`."""
 
-__all__ = ["plot_potential_contours", "plot_density_contours"]
+__all__ = [
+    # potential
+    "plot_potential_contours",
+    "plot_density_contours",
+]
 
-from .contours import plot_density_contours, plot_potential_contours
+from .potential import plot_density_contours, plot_potential_contours
