+++ src/galax/_interop/matplotlib/potential.py
